feat: add options.useMasterKey param for most RestAPI wrappers
also split utils.js to av.js/init.js/utils.js
